+++ CHANGELOG.md
@@ -31,6 +31,7 @@ For a full diff see [`fa9c564...master`][fa9c564...master].
 * Added `@template` annotations to assist with static code analysis ([#128]), by [@localheinz]
 * Removed the fluent interface from `FixtureFactory::defineEntity()` ([#131]), by [@localheinz]
 * Extracted `FieldDefinition\Reference` ([#157]), by [@localheinz]
+* Extracted `FieldDefinition\References` ([#159]), by [@localheinz]
 
 ### Fixed
 
@@ -84,5 +85,6 @@ For a full diff see [`fa9c564...master`][fa9c564...master].
 [#149]: https://github.com/ergebnis/factory-bot/pull/149
 [#155]: https://github.com/ergebnis/factory-bot/pull/155
 [#157]: https://github.com/ergebnis/factory-bot/pull/157
+[#159]: https://github.com/ergebnis/factory-bot/pull/159
 
 [@localheinz]: https://github.com/localheinz